source/app-desktop/requirements-desktop.txt
pywebview
pyinstaller
pillow
